[0698] Fixed PR feedbacks
